+++ kshell/src/main/java/sparklin/kshell/Shared.java
@@ -1,5 +1,0 @@
-package sparklin.kshell;
-
-public class Shared {
-    public static Object __res = null;
-}