+++ sparklin-spark-1.x/src/main/java/sparklin/spark1x/Shared.java
@@ -3,7 +3,7 @@ package sparklin.spark1x;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.sql.SQLContext;
 
-public class Shared extends sparklin.kshell.Shared {
+public class Shared {
     public static JavaSparkContext sc = null;
 
     public static SQLContext sqlContext = null;