+++ sparklin-spark-2.x/src/main/java/sparklin/spark2x/Shared.java
@@ -3,7 +3,7 @@ package sparklin.spark2x;
 import org.apache.spark.api.java.JavaSparkContext;
 import org.apache.spark.sql.SparkSession;
 
-public class Shared extends sparklin.kshell.Shared {
+public class Shared {
     public static JavaSparkContext sc = null;
 
     public static SparkSession spark = null;